Convert AddSubscription components into Typescript
