+++ packages/client/package.json
@@ -106,6 +106,7 @@
     "@types/react-navigation": "^2.0.11",
     "@types/react-redux": "^6.0.4",
     "@types/react-router-dom": "^4.2.7",
+    "@types/react-stripe-elements": "^1.1.7",
     "@types/webpack-env": "^1.13.6",
     "apollo-utilities": "^1.0.3",
     "autoprefixer": "^8.3.0",
