+++ packages/client/src/modules/payments/stripe/recurring/components/AddSubscriptionView.tsx
@@ -1,62 +1,47 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { Elements } from 'react-stripe-elements';
 
 import translate from '../../../../../i18n';
-import { LayoutCenter, clientOnly } from '../../../../common/components';
+import { LayoutCenter, clientOnly } from '../../../../common/components/index.web';
 import { PageLayout } from '../../../../common/components/web';
-import SubscriptionCardForm from './SubscriptionCardForm';
+import SubscriptionCardForm from './SubscriptionCardForm.web';
 import settings from '../../../../../../../../settings';
 
 const ElementsClientOnly = clientOnly(Elements);
 
-class SubscriptionView extends React.Component {
-  static propTypes = {
-    subscribe: PropTypes.func.isRequired,
-    t: PropTypes.func
-  };
-
-  onSubmit = subscribe => async values => {
-    const result = await subscribe(values);
-    const { t } = this.props;
-
-    if (result.errors) {
-      let submitError = {
-        _error: t('errorMsg')
-      };
-      result.errors.map(error => (submitError[error.field] = error.message));
-      throw submitError;
-    }
-  };
-
-  render() {
-    const { subscribe, t } = this.props;
-
-    const renderMetaData = () => (
-      <Helmet
-        title={`${settings.app.name} - ${t('title')}`}
-        meta={[
-          {
-            name: 'description',
-            content: `${settings.app.name} - ${t('meta')}`
-          }
-        ]}
-      />
-    );
-
-    return (
-      <PageLayout>
-        {renderMetaData()}
-        <LayoutCenter>
-          <h1 className="text-center">{t('subTitle')}</h1>
-          <ElementsClientOnly>
-            <SubscriptionCardForm onSubmit={this.onSubmit(subscribe)} action={t('action')} />
-          </ElementsClientOnly>
-        </LayoutCenter>
-      </PageLayout>
-    );
-  }
+// TODO: write types
+interface SubscriptionViewProps {
+  onSubmit: () => void;
+  t: any;
 }
+
+const SubscriptionView = (props: SubscriptionViewProps) => {
+  const { onSubmit, t } = props;
+
+  const renderMetaData = () => (
+    <Helmet
+      title={`${settings.app.name} - ${t('title')}`}
+      meta={[
+        {
+          name: 'description',
+          content: `${settings.app.name} - ${t('meta')}`
+        }
+      ]}
+    />
+  );
+
+  return (
+    <PageLayout>
+      {renderMetaData()}
+      <LayoutCenter>
+        <h1 className="text-center">{t('subTitle')}</h1>
+        <ElementsClientOnly>
+          <SubscriptionCardForm onSubmit={onSubmit} action={t('action')} />
+        </ElementsClientOnly>
+      </LayoutCenter>
+    </PageLayout>
+  );
+};
 
 export default translate('subscription')(SubscriptionView);
